# ongaki-cloud/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Ongaki Cloud API
  version: 1.0.0
servers:
  - url: https://api.ongaki.cloud/v1
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
  /tenants:
    post:
      summary: Create tenant account
      operationId: createTenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
  /projects:
    get:
      summary: List projects
      operationId: listProjects
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
    post:
      summary: Create project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProjectRequest'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Ongaki-API-Key
...
  schemas:
    CreateTenantRequest:
      type: object
      required: [name, slug, ownerEmail]
      properties:
        name:
          type: string
          example: Acme Ltd
        slug:
          type: string
          example: acme
        ownerEmail:
          type: string
          format: email
    Tenant:
      type: object
      required: [id, slug, name, status]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        status:
          type: string
    CreateProjectRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        framework:
          type: string
          default: auto
        repositoryUrl:
          type: string
    Project:
      type: object
      required: [id, tenantId, slug, name]
      properties:
        id:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        framework:
          type: string
...